Added DifficultyAbility.fs
